+++ tests/test_geometry.py
@@ -390,25 +390,25 @@ def test_move_mesh():
 
 
 @pytest.mark.parametrize(
-    "type, expected_tolerance", [("poisson", 1e-12), ("eikonal", 5e-2)]
+    "method, expected_tolerance", [("poisson", 0.28), ("eikonal", 5e-2)]
 )
-def test_boundary_distance(type, expected_tolerance):
+def test_boundary_distance(method, expected_tolerance):
     mesh, _, boundaries, _, _, _ = cashocs.regular_mesh(16)
     dist = cashocs.geometry.compute_boundary_distance(
-        mesh, boundaries=boundaries, boundary_idcs=[1, 2, 3, 4], type=type
+        mesh, boundaries=boundaries, boundary_idcs=[1, 2, 3, 4], method=method
     )
     assert dist.vector().min() >= 0.0
-    assert (dist.vector().max() - 0.5) / 0.5 <= expected_tolerance
+    assert np.abs(dist.vector().max() - 0.5) / 0.5 <= expected_tolerance
 
     dist = cashocs.geometry.compute_boundary_distance(
-        mesh, boundaries=boundaries, boundary_idcs=[1], type=type
+        mesh, boundaries=boundaries, boundary_idcs=[1], method=method
     )
     assert dist.vector().min() >= 0.0
-    assert (dist.vector().max() - 1.0) / 1.0 <= expected_tolerance
+    assert np.abs(dist.vector().max() - 1.0) / 1.0 <= expected_tolerance
 
-    dist = cashocs.geometry.compute_boundary_distance(mesh, type=type)
+    dist = cashocs.geometry.compute_boundary_distance(mesh, method=method)
     assert dist.vector().min() >= 0.0
-    assert (dist.vector().max() - 0.5) / 0.5 <= expected_tolerance
+    assert np.abs(dist.vector().max() - 0.5) / 0.5 <= expected_tolerance
 
 
 def test_named_mesh_import():