+++ CHANGELOG.rst
@@ -13,7 +13,13 @@ of the maintenance releases, please take a look at
 
 * Add a pseudo time stepping solver for nonlinear equations. This uses PETSc's TS and is implemented in :py:func:`cashocs.ts_pseudo_solve`.
 
-* Add a Poisson approach for computing the distance to the wall. This should be more robust than the previous approach of solving the Eikonal equation and is, thus, the new default method.
+* Add a Poisson approach for computing the distance to the wall. This should be more robust than the previous approach of solving the Eikonal equation but less robust. The method can be changed in the configuration (see "New configuration file parameters" below).
+
+* New configuration file parameters:
+
+  * Section ShapeGradient
+
+    * :ini:`distance_method` specifies which method should be used to compute the distance to the boundary. The default is :ini:`distance_method = eikonal`, but the new Poisson approach can be used with :ini:`distance_method = poisson`.
 
 2.3.0 (November 12, 2024)
 -------------------------